% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub_fun_AIC_selection.R
\name{AICselection}
\alias{AICselection}
\title{AICselection}
\usage{
AICselection(
  LL,
  npar,
  n,
  R2 = NULL,
  mnames1 = legend.nm,
  include_marginal = FALSE,
  type2 = 2,
  covnm = NULL,
  rsType = NULL,
  hypoth = NULL,
  LnDet = NULL,
  CE = NULL
)
}
\arguments{
\item{rsType}{default = NULL}

\item{LnDet}{default = NULL}

\item{CE}{Climate enhanced model? default = NULL
list of input data from makeDat() function: "parameters" "rs_dat"     "maplist"    "estparams"  "phases"}

\item{data}{data.frame with models = rows and the following data for each model (columns):
* data$LL  vector of logliklihoods of each model
* data$npar number of parameteris in a given model
* data$n number of data observations (nobs)
* data$mnames1 names for each model
* data$type2 number 1= AIC;2 = AICc ; default = NULL
* data$covnm default = NULL}
}
\value{
a dataframe of summary statistics for AICc of submodels
}
\description{
AICselection runs AIC information criterion analysis on resulting set of recruitment models
}
