% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_load_functions.R
\docType{data}
\name{dirlist}
\alias{dirlist}
\title{load_functions.R}
\format{
An object of class \code{character} of length 26.
}
\usage{
dirlist
}
\description{
A script to load all the functions in the subfolder
}
\details{
----------------------------------------
load_functions.R
subset of futR
kirstin.holsman@noaa.gov
updated 2024
----------------------------------------
}
\keyword{datasets}
