% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub_fun_estRec.R
\name{estRec}
\alias{estRec}
\title{estRec estimate recruitment with options for phasing}
\usage{
estRec(
  version = "futR",
  datlistIN = datlist,
  returnAll = FALSE,
  quiet = FALSE,
  REcompile = TRUE
)
}
\arguments{
\item{datlistIN}{list of input data from makeDat() function: "parameters" "rs_dat"     "maplist"    "estparams"  "phases"}

\item{returnAll}{return all phases? TRUE/FALSE}

\item{quiet}{print out phases}
}
\value{
a list with final (final phase) dlistfinal, dlist (all phases), phase
}
\description{
This function prepares the ceattle estimated data for futR
For more information contact author Kirstin Holsman (kirstin.holsman@noaa.gov)
}
\examples{
mod<-estRec(dataINUSE  =  dataIN_admb)

}
