% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub_fun_get_cov_shape.R
\name{get_cov_shape}
\alias{get_cov_shape}
\title{get_cov_shape.R}
\usage{
get_cov_shape(
  AIC_summryIN,
  s = 1,
  SD_range = 3,
  simulateIN = F,
  sim_nitrIN = 200,
  futR_path = file.path("..", "..", "futR")
)
}
\arguments{
\item{AIC_summryIN}{summary table input}

\item{s}{species number; default = 1}

\item{SD_range}{; default = 3}

\item{simulateIN}{T/F default = F}

\item{sim_nitrIN}{number of iterations for the simulation,
used if simulateIN == T, default = 200}

\item{futR_path}{local file path for the futR directory}
}
\value{
List with plot "p" and a profile of covars
}
\description{
Get the shape of the covariate effects
}
\examples{
#get_cov_shape()

}
