% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub_fun_get_interaction_mat.R
\name{get_interaction_mat}
\alias{get_interaction_mat}
\title{get_interaction_mat.R}
\usage{
get_interaction_mat(covIN, maxIN, ADDTEMP2 = TRUE, cor_cutoff = 0.5)
}
\arguments{
\item{covIN}{covars (matrix with cols for each covar) e.g., pre_spawning_covs[,-1]}

\item{maxIN}{max number of interactions}

\item{ADDTEMP2}{add a temperature ^2 term to those covars with "temp" in the name}

\item{cor_cutoff}{cut off value for correlation matrix (e.g., 0.5), those with cor > than this value will be excluded
get full set of 4 interactions (max_interactions)
returns a list with data.frames of all possible combinations}
}
\description{
Function to get the interaction matrix for env covariates
}
\examples{
get_interaction_mat()

}
