% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub_fun_get_model.R
\name{get_model}
\alias{get_model}
\title{get_model.R}
\usage{
get_model(
  mod,
  recIN,
  simulateIN = FALSE,
  sim_nitrIN = 1000,
  futR_fldr = NULL,
  phasesIN = phases
)
}
\arguments{
\item{mod}{The model name}

\item{recIN}{recruitment values input}

\item{simulateIN}{T/F simulate random noise? default = FALSE,}

\item{sim_nitrIN}{if simulateIN == T, number of random draws, default =  1000,}

\item{futR_fldr}{local path for tht futR directory, default =   NULL,}

\item{phasesIN}{number of phases default =  phases}
}
\value{
List with two objects : lobject and dlist
}
\description{
get the model from the meta evaluation
}
\examples{
get_model()

}
