% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub_fun_getFit.R
\name{getFit}
\alias{getFit}
\title{extract the model fits from the runmod() results}
\usage{
getFit(mm, nm = "recType = 2")
}
\arguments{
\item{mm}{output from the model}

\item{nm}{model name to assign to the model}
}
\value{
maplist a list with map values
}
\description{
This function prepares data for plot_rs()
For more information contact author Kirstin Holsman (kirstin.holsman at noaa.gov)
}
\examples{
 #add


}
