% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub_fun_makeDat.R
\name{makeDat}
\alias{makeDat}
\title{makeDat set up data list and map for futR}
\usage{
makeDat(
  estMode = 1,
  rec_years,
  Rec,
  SSB,
  sdSSB = NULL,
  sdRec = NULL,
  rectype,
  tauIN = 0,
  gammaIN = NULL,
  estparams = c(log_a = TRUE, log_b = TRUE, gamma = FALSE, beta = FALSE, lambda = TRUE,
    epsi_s = FALSE, logsigma = TRUE),
  covars = NULL,
  covars_sd = NULL,
  beta_0 = NULL,
  lambda_0 = NULL,
  startVal = NULL,
  phases = NULL,
  fityrs = NULL,
  REcompile = TRUE,
  sigMethod = 1
)
}
\arguments{
\item{estMode}{estimation mode, 0 = don't est, 1 = estimate par}

\item{rec_years}{required 'years' that match SSB and Rec vectors}

\item{Rec}{required vector of Rec (usually previous yr if Rec = age 1)}

\item{SSB}{required vector of SSB (usually previous yr if Rec = age 1)}

\item{sdSSB}{optional sd of SSB vector a}

\item{sdRec}{optional sd of rec vector}

\item{rectype}{default is 3; Recruitment model formulation based Climate enhanced Rec~SSB functions from Holsman et al. 2019; 
1 = Linear with biomass (y-1), 
2 =  linear, 
3 = Beverton holt, 
4 = Ricker}

\item{estparams}{vector of TRUE / FALSE for each parm,  if parm will be estimated (default is init vals) c(log_a = TRUE, log_b=TRUE, rs_parm=FALSE, logsigma=TRUE)}

\item{covars}{default is NULL; data.frame of covariates (rows) of z-score scaled environmental covariates to evaluate in Rec for each year (columns). Must match  Rec and SSB row index}

\item{covars_sd}{default is NULL; data.frame of sd for covariates (rows) of z-score scaled environmental covariates to evaluate in Rec for each year (columns). Must match Rec and SSB row index}

\item{beta_0}{= NULL,}

\item{lambda_0}{= NULL}

\item{startVal}{default is NULL; starting values for parameters can be a subset list but names must match}

\item{phases}{default is 1 for all parms except sigma (phase 2), but can be specified.}

\item{fityrs}{default is NULL; which years to fit (can be all or subset for out of sample analyses)}

\item{REcompile}{default is TRUE; if true will recompile futR.cpp in TMB TRUE/FALSE}

\item{sigMethod}{sigma R method}
}
\value{
list with :
*  parameters      list of parameters and starting vlaues
*  rs_dat          data.frame data.frame of 'years', 'SSB', 'R_obs' and optionally, 'sdSSB','sdR_obs'
*  maplist         list with map values for calling tmb
*  estparams       vecctor of TRUE/ FALSE for each parameter 
*  phases          vector of phases for each parameter
*  inputs          List of inputs to makeDat()
}
\description{
This function prepares the ceattle estimated data for futR
For more information contact author Kirstin Holsman (kirstin.holsman@noaa.gov)
}
\examples{
 makeDat()
}
