% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub_fun_makefutR_data.R
\name{makefutR_data}
\alias{makefutR_data}
\title{makefutR_data}
\usage{
makefutR_data(fn = "data/in/futR_Inputs.xlsx", export_all = F)
}
\arguments{
\item{fn}{path to data input sheet; default is 'data/in/futR_Inputs.xlsx'}

\item{export_all}{export all of the inputs from the xls file}
}
\value{
datlist  list of input data and parameters used for the runmod() dependency
\itemize{
\item datlist$parameters  list of parameters for the TMB futR model:
[] datlist$parameters$log_a    scalar starting value for log_a (intercept)
[] datlist$parameters$log_b    scalar starting value for log_b (slope)
[] datlist$parameters$beta     vector of starting values for lambda (env. effects on pre-spawning)
[] datlist$parameters$lambda   vector of starting values for beta (env. effects on post-spawning)
[] datlist$parameters$epsi_s   starting value for epsi_s (error around S estimates)
[] datlist$parameters$logsigma starting value for logsigma (process error)
[] datlist$parameters$skipFit  starting value for skipFit (switch used for projecting)
\item datlist$rs_dat list of data to read into the model
--- Observation Error inputs
[] tau        scalar for observation and process error
[] sigMethod  Integer (switch): which method for estimating sigma
--- Covariate inputs
[] ncov       Integer number of environmental covariates
[] cov_type   (defunct) vector of covariate type if 1 then run ration
[] nyrs       Integer number of years in the data
[] lambda_0   matrix (ncov, nyrs), binary (0,1) matrix of covar. inclusion in post-spawning effects
[] beta_0     matrix (ncov, nyrs), binary (0,1) matrix of covar. inclusion in pre-spawning effects
[] sdrs_cov   matrix (ncov, nyrs) of the stdev of environmental covariates
[] rs_cov     matrix (ncov, nyrs) of the of environmental covariates
--- RS data inputs
[] sdR        vector (nyrs) of stdev of recruitment in a given year
[] sdS        vector (nyrs) of stdev of spawnerindex in a given year
[] R_obs      vector (nyrs) of recruitment in a given year
[] S_obs      vector (nyrs) of spawner index in a given year
[] years      vector (nyrs) recruiment year
--- Switches
[] rectype    Integer (switch) which type of recruitment model: 1 = loglinear, 2 = loglinear ~f(S), 3 = Beverton-holt, 4 = Ricker
[] estMode    Integer (switch) should the model run in estimation mode (y=1,no=0)?
\item datlist$maplist   list of the map for the TMB model
\item datlist$estparams vector of T/F of which parameters will be estimated
\item datlist$phases    (defunct) phases for estimating parameters
\item datlist$inputs    archive of intputs for this function
}
}
\description{
reads an xlxs input file and creates the datlist for input into runmod()
For more information contact author Kirstin Holsman (kirstin.holsman at noaa.gov)
}
\examples{
#add
}
