% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub_fun_makeMap.R
\name{makeMap}
\alias{makeMap}
\title{makeMap.R}
\usage{
makeMap(param = parameters, estpar = estparams)
}
\arguments{
\item{estpar}{vector of T/F of which parameters will be estimated}

\item{parameters}{list of starting values and parameters for the TMB futR() model}
}
\description{
makeMap will create a map list for reading into the runmod() function
}
\details{
This function prepares the estimated data for futR
This function is a subset of makeDat()
For more information contact author Kirstin Holsman (kirstin.holsman at noaa.gov)
}
\examples{

datlist   <- readMake_futR_data("data/in/futR_Inputs.xlsx" )

estparams <- datlist$estparams
estparams

parameters <- atlist$parameters
parameters
maplist <- makeMap(param=parameters,estpar=estparams)
}
