% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub_fun_plot_par_pdf.R
\name{plot_par_pdf}
\alias{plot_par_pdf}
\title{plot_par.R}
\usage{
plot_par_pdf(df = df1_t0)
}
\arguments{
\item{df}{dataframe of outputs from runmod(), e.g., mm <- runmod(...)
\itemize{
\item model     model name to be displayed on the legend
\item estimate  as.vector(mm$sim)
\item parameter names( mm$mle)\link{row(mm$sim)}
}}
}
\value{
list
}
\description{
plot the pdf of each parameter estimated in the model
}
\details{
For more information contact author Kirstin Holsman (kirstin.holsman at noaa.gov)
}
\examples{
# read in the data

#datlist <- datlist2 <-  
#readMake_futR_data("data/in/futR_Inputs.xlsx" )

# set the sigMethod to 1
#now compare plots
# plot_par_pdf(rbind(dfR4_s1,dfR4_s3))
}
