% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub_fun_predict_futR.R
\name{predict_futR}
\alias{predict_futR}
\title{predict_futR.R}
\usage{
predict_futR(
  modelIN = mm,
  newdata,
  src_fldr = "src/TMB",
  futR_fldr = NULL,
  newpar
)
}
\arguments{
\item{modelIN}{model output from runRecMod()}

\item{newdata}{new predictive data to simulate (can be new S_obs or cov)}

\item{src_fldr}{src folder path}

\item{futR_fldr}{NULL}

\item{newpar}{new parameters to run through the mode description}
}
\value{
pred_model$report()
}
\description{
Predict recruitment from new data and or new par
}
