% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub_fun_profile_covars.R
\name{profile_covars}
\alias{profile_covars}
\title{profile_covars.R}
\usage{
profile_covars(
  cov_nm = "Spring_temp_surface5m",
  T2 = T,
  SD_range = 3,
  sim_nitrIN = 100,
  simulateIN = TRUE,
  modIN,
  recIN,
  phasesIN
)
}
\arguments{
\item{cov_nm}{covariate name}

\item{T2}{T/F squared term for the parameter}

\item{SD_range}{range of SD}

\item{sim_nitrIN}{number of random simulations}

\item{simulateIN}{simulate random iterations sim_nitrIN}

\item{modIN}{rec model e.g., AIC_summry[\link{s}]$topRicker_R2}

\item{recIN}{recruitment fit from the rec model, e.g., AIC_summry[\link{s}]$topRicker_R2_rec_fit}

\item{phasesIN}{phases for each parm}
}
\value{
list(sim_df=sim_df,R_hat_sim=R_hat_sim, poly=aa,main = bb,p=p)
}
\description{
profile the covariates to get partial effects
}
