% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub_fun_profile_SSB.R
\name{profile_ssb}
\alias{profile_ssb}
\title{profile the SSB to get partial effects}
\usage{
profile_ssb(
  cov_nm_list = c("Spring_temp_surface5m", "Summer_temp_surface5m",
    "Winter_temp_bottom5m"),
  nameIN = "hindcast 2020",
  newdataIN = hind \%>\% filter(year == 2020) \%>\% data.frame(),
  T2 = T,
  SD_range = 7,
  steps = 100,
  sim_nitrIN = 100,
  simulateIN = TRUE,
  hind_yearsIN = 1980:2023,
  modIN = AIC_summry[[s]]$topRicker_R2,
  recIN = AIC_summry[[s]]$topRicker_R2_rec_fit,
  phasesIN = phases
)
}
\arguments{
\item{cov_nm_list}{= c("Spring_temp_surface5m","Summer_temp_surface5m","Winter_temp_bottom5m"),}

\item{nameIN}{= "hindcast 2020",}

\item{newdataIN}{= hind%>%filter(year==2020)%>%data.frame(),}

\item{T2}{= T,}

\item{SD_range}{= 7,}

\item{steps}{= 100,}

\item{sim_nitrIN}{= 100,}

\item{simulateIN}{= TRUE,}

\item{hind_yearsIN}{= 1980:2023,}

\item{modIN}{= AIC_summry[[s]]$topRicker_R2,}

\item{recIN}{= AIC_summry[[s]]$topRicker_R2_rec_fit,}

\item{phasesIN}{= phases description}
}
\description{
profile the SSB to get partial effects
}
