% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rec}
\alias{rec}
\title{rec}
\format{
A dataframe with the following columns
\describe{
\item{rec_year}{ year of recruitment}
\item{SSB}{Spawning biomass in the previous year}
\item{Robs}{Recruitment in the year}
\item{sdSSB}{Error in the Spawning biomass estimate}
\item{sdRobs}{Error in the recruitment estimate}
\item{fit}{0/1 no/yes fit in the model?}
}
}
\source{
K. Holsman et al. 2024
}
\usage{
rec
}
\description{
A data frame with recruitment
}
\keyword{datasets}
