% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub_fun_recompile.R
\name{recompile}
\alias{recompile}
\title{recompile()}
\usage{
recompile(model = "futR")
}
\arguments{
\item{model}{(without the '.cpp'); default is "futR"}
}
\value{
list
}
\description{
recompiles the model
}
\details{
For more information contact author Kirstin Holsman (kirstin.holsman@noaa.gov)
}
\examples{
recompile(model="futR")
}
