% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub_fun_rmd2md.R
\name{rmd2md}
\alias{rmd2md}
\title{rmd2md()}
\usage{
rmd2md(rmd_fl, md_fl = "README")
}
\arguments{
\item{rmd_fl}{name of the markdown file (without the .Rmd)}

\item{md_fl}{name of the output; default is 'README'}
}
\description{
convert markdown doc to readme file for github
}
\details{
For more information contact author Kirstin Holsman (kirstin.holsman at noaa.gov)
}
\examples{
# convert the mymarkdown.Rmd to README.md
#rmd2md(rmd_fl="mymarkdown",md_fl = "README")  
}
