% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub_fun_runRecMod.R
\name{runRecMod}
\alias{runRecMod}
\title{run the futR() recruitment model}
\usage{
runRecMod(
  dlistIN,
  src_fldr = "src",
  version = "futR",
  silentIN = TRUE,
  se.fit = TRUE,
  simulate = FALSE,
  sim_nitr = 1000,
  recompile = FALSE,
  maxitr = 10000,
  maxeval = 10000
)
}
\arguments{
\item{dlistIN}{list of input data and parameters used for the makeADfun() dependency
* dlistIN$parameters  list of parameters for the TMB futR model:
   [] dlistIN$parameters$log_a    scalar starting value for log_a (intercept)
   [] dlistIN$parameters$log_b    scalar starting value for log_b (slope) 
   [] dlistIN$parameters$beta     vector of starting values for lambda (env. effects on pre-spawning)   
   [] dlistIN$parameters$lambda   vector of starting values for beta (env. effects on post-spawning)   
   [] dlistIN$parameters$epsi_s   starting value for epsi_s (error around S estimates) 
   [] dlistIN$parameters$logsigma starting value for logsigma (process error)
   [] dlistIN$parameters$skipFit  starting value for skipFit (switch used for projecting)
* dlistIN$rs_dat list of data to read into the model
   --- Observation Error inputs
   [] tau        scalar for observation and process error
   [] sigMethod  Integer (switch): which method for estimating sigma
   --- Covariate inputs
   [] ncov       Integer number of environmental covariates
   [] cov_type   (defunct) vector of covariate type if 1 then run ration
   [] nyrs       Integer number of years in the data
   [] lambda_0   matrix (ncov, nyrs), binary (0,1) matrix of covar. inclusion in post-spawning effects
   [] beta_0     matrix (ncov, nyrs), binary (0,1) matrix of covar. inclusion in pre-spawning effects
   [] sdrs_cov   matrix (ncov, nyrs) of the stdev of environmental covariates
   [] rs_cov     matrix (ncov, nyrs) of the of environmental covariates
   --- RS data inputs
   [] sdR        vector (nyrs) of stdev of recruitment in a given year
   [] sdS        vector (nyrs) of stdev of spawnerindex in a given year
   [] R_obs      vector (nyrs) of recruitment in a given year
   [] S_obs      vector (nyrs) of spawner index in a given year
   [] years      vector (nyrs) recruiment year
   --- Switches
   [] rectype    Integer (switch) which type of recruitment model: 1 = loglinear, 2 = loglinear ~f(S), 3 = Beverton-holt, 4 = Ricker
   [] estMode    Integer (switch) should the model run in estimation mode (y=1,no=0)?
* dlistIN$maplist   list of the map for the TMB model
* dlistIN$estparams vector of T/F of which parameters will be estimated
* dlistIN$phases    (defunct) phases for estimating parameters
* dlistIN$inputs    archive of intputs for this function}

\item{src_fldr}{subfolder where model is located. Default is 'src'}

\item{version}{recruitment model to run. Default is 'futR'}

\item{silentIN}{T/F , default = 'TRUE'}

\item{se.fit}{T/F , default = 'TRUE'}

\item{simulate}{T/F , simulate the model? default = 'FALSE'}

\item{sim_nitr}{number of iterations for the simulate function}

\item{recompile}{recompile the .cpp file? default is 'FALSE'}

\item{maxitr}{10000  Max iterations for fitting the objective function}

\item{maxeval}{10000  Max evaluations for fitting the objective function}
}
\value{
returns  summary of the model including the mle
}
\description{
runRecMod() will run the futR() recruitment model
For more information contact author Kirstin Holsman (kirstin.holsman@noaa.gov)
}
\examples{
datlist <- readMake_futR_data("data/in/futR_Inputs.xlsx" )
mm      <- runRecMod(dlistIN   = datlist, version   = 'futR',recompile = FALSE,simulate  = TRUE,sim_nitr  = 1000)  
}
