% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{testTHIS}
\alias{testTHIS}
\title{profile the SSB to get partial effects}
\usage{
testTHIS(msg)
}
\arguments{
\item{msg}{}
}
\value{
list of ssb profiles
}
\description{
profile the SSB to get partial effects
}
