% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub_fun_THEMES_GGPLOT.R
\name{theme_kir}
\alias{theme_kir}
\title{Kir's ggplot2 go-to theme based on theme_hrbrmstr}
\usage{
theme_kir(
  base_family = "Hind",
  base_size = 11,
  strip_text_family = "Hind Medium",
  strip_text_size = 12,
  plot_title_family = "Hind",
  plot_title_size = 18,
  plot_title_margin = 10,
  plot_title_just = "r",
  xangle = 0,
  plot_margin = margin(30, 30, 30, 30),
  subtitle_family = "Hind Medium",
  subtitle_size = 12,
  subtitle_face = "plain",
  subtitle_margin = 15,
  panel_face = "plain",
  caption_family = "Hind Light",
  caption_size = 9,
  caption_margin = 10,
  axis_title_family = subtitle_family,
  axis_title_size = 9,
  axis_title_just = "cm",
  xaxis_size = 9,
  yaxis_size = 9,
  grid = TRUE,
  axis = FALSE,
  sub_title_size = 10,
  sub_title_col = "black",
  sub_title_just = "c",
  panel_dist_y = 2,
  panel_dist_x = 2,
  ticks = FALSE
)
}
\arguments{
\item{base_family}{base font family}

\item{base_size}{base font size}

\item{strip_text_family}{facet label font family}

\item{strip_text_size}{facet label text size}

\item{plot_title_family}{plot tilte family}

\item{plot_title_size}{plot title font size}

\item{plot_title_margin}{plot title margin}

\item{plot_margin}{plot margin}

\item{subtitle_family}{plot subtitle family}

\item{subtitle_size}{plot subtitle size}

\item{subtitle_margin}{plot subtitle margin}

\item{caption_family}{plot caption family}

\item{caption_size}{plot caption size}

\item{caption_margin}{plot caption margin}

\item{axis_title_family}{axis title font family}

\item{axis_title_size}{axis title font size}

\item{axis_title_just}{axis title font justification \code{blmcrt} rt = right top, bl = bottom left, etc.}

\item{grid}{panel grid (\code{TRUE}, \code{FALSE}, or a combination of
\code{X}, \code{x}, \code{Y}, \code{y})}

\item{axis}{axis \code{TRUE}, \code{FALSE}, [\code{xy}]}

\item{ticks}{ticks}

\item{subplottxt}{color of the subplot titles}
}
\description{
It requires installing Hind fonts unless you change the font parameters
}
