% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub_fun_vec2list.R
\name{vec2list}
\alias{vec2list}
\title{vec2list()}
\usage{
vec2list(x)
}
\arguments{
\item{x}{Numeric vector.}
}
\value{
list
}
\description{
convert vector to list using names
}
\examples{
bmi.vals <- rnorm(n = 50, mean = 25, sd = 3)
vec2list(bmi.vals)

}
