% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub_fun_AIC_selection.R
\name{AICselection}
\alias{AICselection}
\title{AICselection}
\usage{
AICselection(
  LL,
  npar,
  n,
  R2 = NULL,
  mnames1 = NULL,
  include_marginal = FALSE,
  type2 = 2,
  covnm = NULL,
  rsType = NULL,
  hypoth = NULL,
  LnDet = NULL,
  CE = NULL
)
}
\arguments{
\item{LL}{vector of logliklihoods of each model}

\item{npar}{number of parameteris in a given model}

\item{n}{number of data observations (nobs)}

\item{R2}{default is NULL, vector of R2 values for each model}

\item{mnames1}{names for each model}

\item{include_marginal}{T/F include mariginal AIC}

\item{type2}{if 1 then AIC is calculated, if 2 then AICc (for small sample sizes is calculated)}

\item{covnm}{default is NULL, covariate num}

\item{rsType}{default is NULL, recruitment type}

\item{hypoth}{default is NULL, hypothesis}

\item{LnDet}{the natural logarithm of the determinant of a matrix  X;  defualt is NULL}

\item{CE}{Climate enhanced model? default = NULL
list of input data from makeDat() function: "parameters" "rs_dat"     "maplist"    "estparams"  "phases"}
}
\value{
a dataframe of summary statistics for AICc of submodels
}
\description{
AICselection runs AIC information criterion analysis on resulting set of recruitment models
}
\examples{
#add

}
