% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub_fun_get_SSB_shape.R
\name{get_SSB_shape}
\alias{get_SSB_shape}
\title{get the shape of the recruitment curve}
\usage{
get_SSB_shape(
  AIC_summryIN,
  simname = "hindcast 2020",
  SD_range = 7,
  modIN,
  recfitIN,
  simulateIN = F,
  sim_nitrIN = 200,
  NewCovsIN,
  steps = 100,
  phases,
  futR_path = "../../futR"
)
}
\arguments{
\item{AIC_summryIN}{AIC summary to input}

\item{simname}{simulation name}

\item{SD_range}{range of the SD}

\item{modIN}{model input, e.g., AIC_summry[[1]]$topRicker_R2}

\item{recfitIN}{recruitment model fit, e.g., AIC_summry[[1]]$topRicker_R2_rec_fit}

\item{simulateIN}{simulate T or F}

\item{sim_nitrIN}{number of simulations}

\item{NewCovsIN}{covariates to input, e.g, hind%>%dplyr::filter(year==2020)%>%data.frame()}

\item{steps}{number of steps}

\item{phases}{Input of phases for each parameter}

\item{futR_path}{= path for the futR model}
}
\value{
The sum of `x` and `y`
}
\description{
This function will get the SSB R shape
For more information contact author Kirstin Holsman (kirstin.holsman at noaa.gov)
}
\examples{
#get_SSB_shape()
}
