% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub_fun_get_SSB_shape.R
\name{get_SSB_shape}
\alias{get_SSB_shape}
\title{get_SSB_shape.R}
\usage{
get_SSB_shape(
  AIC_summryIN,
  simname = "hindcast 2020",
  SD_range = 7,
  modIN = AIC_summry[[1]]$topRicker_R2,
  recfitIN = AIC_summry[[1]]$topRicker_R2_rec_fit,
  simulateIN = F,
  sim_nitrIN = 200,
  NewCovsIN = hind \%>\% filter(year == 2020) \%>\% data.frame(),
  steps = 100,
  futR_path = "../../futR"
)
}
\arguments{
\item{AIC_summryIN}{,}

\item{simname}{= "hindcast 2020",}

\item{SD_range}{= 7,}

\item{modIN}{= AIC_summry[[1]]$topRicker_R2,}

\item{recfitIN}{= AIC_summry[[1]]$topRicker_R2_rec_fit,}

\item{sim_nitrIN}{= 200,}

\item{NewCovsIN}{=hind%>%filter(year==2020)%>%data.frame(),}

\item{steps}{= 100,}

\item{futR_path}{= "../../futR")}

\item{x}{A description of the parameter 'x'. The
description can span multiple lines.}

\item{y}{A description of the parameter 'y'}

\item{simulateIN=}{F,}
}
\value{
The sum of `x` and `y`.
}
\description{
get the SSB / R shape
}
\examples{
get_SSB_shape()
}
