% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub_fun_plot_rs.R
\name{plot_rs}
\alias{plot_rs}
\title{plot_rs}
\usage{
plot_rs(rec_df, cbp = "Dark2", obscol = "gray")
}
\arguments{
\item{rec_df}{dataframe of recruitment and spawners generated by getFit(), column headers include
* R_obs (recruitment)
* S_obs (Spawning biomass)
* R_hat (fitted Recruitment)
* S_hat (predicted spawning biomass )
* year (recruitment year)
* model (factor of rec model names)}

\item{cbp}{color_brewer palette for the different models; default = "Dark2"}

\item{obscol}{color for the "observed" recruitment; default = "gray"}
}
\value{
list
}
\description{
plot the recruitment model in two panels a) R~f(S) and b) R~f(Year)
For more information contact author Kirstin Holsman (kirstin.holsman at noaa.gov)
}
\examples{
# add
}
