% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub_fun_profile_covars.R
\name{profile_covars}
\alias{profile_covars}
\title{profile_covars.R}
\usage{
profile_covars(
  cov_nm = "Spring_temp_surface5m",
  T2 = T,
  SD_range = 3,
  sim_nitrIN = 100,
  simulateIN = TRUE,
  modIN = AIC_summry[[s]]$topRicker_R2,
  recIN = AIC_summry[[s]]$topRicker_R2_rec_fit,
  phasesIN = phases
)
}
\arguments{
\item{cov_nm}{= "Spring_temp_surface5m",}

\item{T2}{= T,}

\item{SD_range}{=  3,}

\item{sim_nitrIN}{= 100,}

\item{simulateIN}{= TRUE,}

\item{modIN}{= AIC_summry[[s]]$topRicker_R2,}

\item{recIN}{= AIC_summry[[s]]$topRicker_R2_rec_fit,}

\item{phasesIN}{= phases description}
}
\value{
list(sim_df=sim_df,R_hat_sim=R_hat_sim, poly=aa,main = bb,p=p)
}
\description{
profile the covariates to get partial effects
}
