% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub_fun_profile_SSB.R
\name{profile_ssb}
\alias{profile_ssb}
\title{profile the SSB to get partial effects}
\usage{
profile_ssb(
  cov_nm_list = c("Spring_temp_surface5m", "Summer_temp_surface5m",
    "Winter_temp_bottom5m"),
  nameIN = "hindcast 2020",
  newdataIN = hind \%>\% filter(year == 2020) \%>\% data.frame(),
  T2 = T,
  SD_range = 7,
  steps = 100,
  sim_nitrIN = 100,
  simulateIN = TRUE,
  hind_yearsIN = 1980:2023,
  modIN = AIC_summry[[s]]$topRicker_R2,
  recIN = AIC_summry[[s]]$topRicker_R2_rec_fit,
  phasesIN = phases
)
}
\arguments{
\item{cov_nm_list}{vector of covariate names character string}

\item{nameIN}{simulation input}

\item{newdataIN}{hindcast data input}

\item{T2}{use temperature squared T or F}

\item{SD_range}{range of SD}

\item{steps}{number of steps}

\item{sim_nitrIN}{number of iterations}

\item{simulateIN}{simulate it T or F}

\item{hind_yearsIN}{years of the hindcast}

\item{modIN}{rec model in}

\item{recIN}{fit recruitment estimates}

\item{phasesIN}{phases}
}
\value{
list of ssb profiles
}
\description{
profile the SSB to get partial effects
}
